// FollowedSites/src/webparts/followedSites/interfaces/IFollowedResult.ts
export default interface IFollowedResult {
  AccountName?: any;
  ActorType: number;
  CanFollow: boolean;
  ContentUri: string;
  EmailAddress?: any;
  FollowedContentUri?: any;
  Id: string;
  ImageUri?: any;
  IsFollowed: boolean;
  LibraryUri?: any;
  Name: string;
  PersonalSiteUri?: any;
  Status: number;
  StatusText?: any;
  TagGuid: string;
  Title?: any;
  Uri: string;
}